requirements.txt
freezegun==0.3.10
python-dateutil==2.6.0
ciscosparkapi==0.7.1
flask-io==1.11.0
flask-restplus==0.9.2
pytest==3.4.2
pytest-cov==2.5.1
numpy==1.12.1
matplotlib==2.0.0
attrdict==2.0.0
APScheduler==2.1.2
Flask-APScheduler==1.8.0